Grade Check 1
Several issues with input and the for loop. The while loop worked, but even when program was functional it continued for all 10 iterations.

Missing a doc string for the document and all of the function definitions.

Need to add EOL comments to ensure fellow developers understand what is occurring within your program. This helps you as well.